+++ docs/CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Added AntiStorage documentation to the book. ([#997])
 * You can now stop the rotation of the FreeRotationSystem by setting HideCursor.hide value to false. ([#997])
 * Support for logging to file, toggle for logging to stdout. ([#976], [#994])
+* Added a `Hidden` Component, that hides a single entity, and a HideHierarchySystem that toggles `Hidden` on all children when used. ([#1001])
 * Documentation for drawing sprites. ([#971])
 * Added `shadow_update()` and `shadow_fixed_update()` to the `State` trait. ([#1006])
 * Added configurable width for debug lines ([#1016])
@@ -105,6 +106,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#981]: https://github.com/amethyst/amethyst/pull/981
 [#994]: https://github.com/amethyst/amethyst/pull/994
 [#997]: https://github.com/amethyst/amethyst/pull/997
+[#1001]: https://github.com/amethyst/amethyst/pull/1001
 [#1006]: https://github.com/amethyst/amethyst/pull/1006
 [#1008]: https://github.com/amethyst/amethyst/pull/1008
 [#1016]: https://github.com/amethyst/amethyst/pull/1016